+++ server/helpers.mjs
@@ -1,3 +1,6 @@
+import {telegram} from "./telegram.mjs";
+import {discord} from "./discord.mjs";
+
 export const parseTelegramChatIDs = s => s ? s.split(",").map( v => v.trim() ) : ""
 
 export const timestamp = () => {
@@ -9,4 +12,16 @@ export const timestamp = () => {
     let M = String(today.getMinutes()).padStart(2, '0');
 
     return `${d}/${m}/${y} ${H}:${M}`;
+}
+
+export const sendAlert = (check, message) => {
+    const {telegramToken, alertToTelegram, telegramChatIDAlert, discordWebHook, alertToDiscord} = globalThis.config
+
+    if (telegramToken && alertToTelegram.includes(check)) {
+        telegram(message, {token: telegramToken, recipients: telegramChatIDAlert})
+    }
+
+    if (discordWebHook && alertToDiscord.includes(check)) {
+        discord(discordWebHook, message)
+    }
 }