+++ server/price-sender.mjs
@@ -2,20 +2,20 @@ import fetch from "node-fetch"
 import {TELEGRAM_BOT_URL} from "./telegram.mjs"
 import {parseTelegramChatIDs, timestamp} from "./helpers.mjs";
 import {discord} from "./discord.mjs";
-import {getPriceInfo} from "./coingecko.mjs";
 
-export const processPriceSend = async (config) => {
-    const {discordWebHook, telegramChatID, telegramToken} = config
-    const {currency = 'usd', interval = 3600000, targets = []} = config.price
+export const processPriceSend = async () => {
+    if (!globalThis.config) return
+
+    const {discordWebHook, telegramChatID, telegramToken, price} = globalThis.config
+    const {currency = 'usd', interval = 3600000, targets = []} = price
     const TELEGRAM_URL = TELEGRAM_BOT_URL.replace("%TOKEN%", telegramToken)
-    let price
 
-    if (!config || targets.length === 0) return
+    if (!targets.length) return
 
-    price = await getPriceInfo(currency)
+    let data = globalThis.priceInfo
 
-    if (price && price.length) {
-        const mina = price[0]
+    if (data && data.length) {
+        const mina = data[0]
         const message = `Current Mina price is ${mina.current_price} ${currency.toUpperCase()}`
         const ids = parseTelegramChatIDs(telegramChatID)
 
@@ -30,6 +30,6 @@ export const processPriceSend = async (config) => {
         }
     }
 
-    setTimeout(() => processPriceSend(config), interval)
+    setTimeout(() => processPriceSend(), interval)
 }